requirements.txt
telethon~=1.38.1
requests~=2.32.3
nltk~=3.9.1
langdetect~=1.0.9
uvicorn~=0.32.1
fastapi~=0.115.6
python-logstash-async~=4.0.1